// src/interfaces/SessionRequest.ts
import {Request} from "express";
import * as session from 'express-session';

/**
 * SessionRequest interface extends the Express Request object to include session data.
 * It includes the serverHash property, which stores the hash generated by the server
 * for authentication purposes.
 *
 * @property {session.Session & Partial<session.SessionData> & { serverHash?: string }} session - The session object containing serverHash.
 */
interface SessionRequest extends Request {
    session: session.Session & Partial<session.SessionData> & { serverHash?: string };
}

export default SessionRequest;